refactor : 구조 변경 및 memberEntity에서 @Setter 삭제
