+++ src/main/java/com/ward/ward_server/api/user/auth/security/JwtIssuer.java
@@ -1,4 +1,4 @@
-package com.ward.ward_server.global.auth.security;
+package com.ward.ward_server.api.user.auth.security;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
@@ -14,7 +14,6 @@ import java.util.List;
 public class JwtIssuer {
     private final JwtProperties properties;
 
-    //TODO 로그인 시 JWT 반환 PAYLOAD 로 username, role, issueAt, expiration 네 가지 항목으로 바꾸기
     public String issue(long userId, String email, List<String> roles) {
         Date issuedAt = new Date();
         Date expiresAt = new Date(System.currentTimeMillis() + Duration.ofDays(1).toMillis());
