+++ src/main/java/com/ward/ward_server/api/user/entity/UserEntity.java
@@ -1,15 +1,12 @@
 package com.ward.ward_server.api.user.entity;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.ward.ward_server.api.user.entity.enumtype.UserRole;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
-@Setter
 @Entity
 @Table(name = "users")
 @NoArgsConstructor(access = AccessLevel.PROTECTED) //JPA 쓰면서 protected 키워드는 생성해서 쓰지말라는 의미
@@ -27,17 +24,27 @@ public class UserEntity {
     @Enumerated(EnumType.STRING)
     private UserRole userRole; //[ROLE_USER,ROLE_ADMIN]
 
-    @JsonIgnore
     private String password;
 
+    private String extraInfo;
+
     //==생성 메서드==//
-    public static UserEntity createUser(String username, String name, String email, String password) {
-        UserEntity userEntity = new UserEntity();
-        userEntity.setUsername(username);
-        userEntity.setName(name);
-        userEntity.setEmail(email);
-        userEntity.setUserRole(UserRole.ROLE_USER);
-        userEntity.setPassword(password);
-        return userEntity;
+//    public static UserEntity createUser(String username, String name, String email, String password) {
+//        UserEntity userEntity = new UserEntity();
+//        userEntity.setUsername(username);
+//        userEntity.setName(name);
+//        userEntity.setEmail(email);
+//        userEntity.setUserRole(UserRole.ROLE_USER);
+//        userEntity.setPassword(password);
+//        return userEntity;
+//    }
+
+    //==생성 메서드==//
+    public UserEntity(String username, String name, String email, String password) {
+        this.username = username;
+        this.name = name;
+        this.email = email;
+        this.userRole = UserRole.ROLE_USER;
+        this.password = password;
     }
 }
