+++ src/test/java/com/ward/ward_server/controller/WithMockUserSecurityContextFactory.java
@@ -1,7 +1,7 @@
 package com.ward.ward_server.controller;
 
-import com.ward.ward_server.global.auth.security.CustomUserDetails;
-import com.ward.ward_server.global.auth.security.UserPrincipalAuthenticationToken;
+import com.ward.ward_server.api.user.auth.security.CustomUserDetails;
+import com.ward.ward_server.api.user.auth.security.UserPrincipalAuthenticationToken;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
